kernelscripts: define CC variable
CAF's script expects this kernel commit to be present:
"scripts: gcc-wrapper: Use wrapper to check compiler warnings".
In some cases, it is reverted or not present, which causes
compile to try to use GCC 4.9 unsuccessfully without this commit.

Since we build with -Werror anyway, this commit is safe because
gcc-wrapper is redundant.  Google also reverts the same CAF commit.

Change-Id: Ie172be870737b7960eb3fc595f51e31e753992ed