Revert "kernel_definitions: Make directory dependencies order-only."
This reverts commit 1b75d060e00bab2553e291aa04e0edc9a28d64b4.

Reason for revert: Switched to a different base for my kernel and it did not rebuild boot.img.  Reverting this change rebuilt.  There may be things we can take from here to improve dirty build time, but as it stands, this is doing more harm than good, especially for early development.

Change-Id: I6bea9936c354aab00bd74ad18a8c5f44eac4b3a9