Update patch set 1
Patch Set 1:

inb4 Alex asks whyyyyyyyy

Elias is working on building for begonia, which has a Mediatek chip.  This isn't too hard to make agnostic, so we should.  At some point, there may be common flags for non-QC chips, but at this point, there are not so maintainers would have to specify in device tree

Patch-set: 1